+++ ft_strlcpy.c
@@ -6,14 +6,14 @@
 /*   By: musakbul <musakbul@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/01 13:07:53 by musakbul          #+#    #+#             */
-/*   Updated: 2025/06/26 16:37:36 by musakbul         ###   ########.fr       */
+/*   Updated: 2025/06/30 17:38:21 by musakbul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, unsigned int size)
 {
 	unsigned int	i;
 	unsigned int	len;
