+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: musakbul <musakbul@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/30 15:22:41 by musakbul          #+#    #+#             */
-/*   Updated: 2025/06/26 19:12:59 by musakbul         ###   ########.fr       */
+/*   Updated: 2025/06/30 17:37:52 by musakbul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	counter;
